# lib/api-zod/src/generated/types/proposalSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sistema de Gerenciamento e Criação de Propostas Comerciais
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalSummaryStatus } from './proposalSummaryStatus';

export interface ProposalSummary {
  id: string;
  status: ProposalSummaryStatus;
  propType: string;
  propMonth: string;
  propYear: string;
  /** @nullable */
  campTag?: string | null;
  /** @nullable */
  clientLine1?: string | null;
  /** @nullable */
  advertiserName?: string | null;
  /** @nullable */
  fromTemplateName?: string | null;
  createdByName?: string;
  createdAt: string;
  updatedAt: string;
}