# lib/api-zod/src/generated/types/proposalTemplateProduct.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sistema de Gerenciamento e Criação de Propostas Comerciais
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalTemplateProductColor } from './proposalTemplateProductColor';

export interface ProposalTemplateProduct {
  id: string;
  order: number;
  qty: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  program?: string | null;
  tags?: string[];
  color: ProposalTemplateProductColor;
}