# lib/api-zod/src/generated/types/statBlock.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sistema de Gerenciamento e Criação de Propostas Comerciais
 * OpenAPI spec version: 0.1.0
 */

export interface StatBlock {
  num: string;
  suf: string;
  desc: string;
}